% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap incidence time series}
\usage{
bootstrap(x, randomise_groups = FALSE)
}
\arguments{
\item{x}{An \code{incidence} object.}

\item{randomise_groups}{A \code{logical} indicating whether groups should be
randomised as well in the resampling procedure; respective group sizes will
be preserved, but this can be used to remove any group-specific temporal
dynamics. If \code{FALSE} (default), data are resampled within groups.}
}
\value{
An \code{incidence} object.
}
\description{
This function can be used to bootstrap \code{incidence} objects. Bootstrapping is
done by sampling with replacement the original input dates. See \code{details} for
more information on how this is implemented.
}
\details{
As original data are not stored in \code{incidence} objects, the
bootstrapping is achieved by multinomial sampling of date bins weighted by
their relative incidence.
}
\examples{

if (require(outbreaks) && require(ggplot2)) {
  i <- incidence(fluH7N9_china_2013$date_of_onset)
  i
  plot(i)

  ## one simple bootstrap
  x <- bootstrap(i)
  x
  plot(x)

  ## use it to find CI for epidemic peak
  find_peak <- function(x) x$dates[which.max(pool(x)$counts)]
  find_peak(i)

  ## peaks on 100 bootstrap samples
  peak_boot <- replicate(100,
                         find_peak(bootstrap(i)),
                         simplify = FALSE)

  ## convert to vector without losing Date class
  peak_boot <- do.call(c, peak_boot)

  ## check distribution of peaking times
  summary(peak_boot)
  plot(i) + geom_density(data = data.frame(peak = peak_boot),
                         aes(x = peak, y = 10 * ..scaled..),
                         alpha = .2, fill = "red", color = "red")

}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
