% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit}
\alias{fit}
\alias{print.incidence.fit}
\title{Fit exponential models to incidence data}
\usage{
fit(x, split = NULL, level = 0.95)

\method{print}{incidence.fit}(x, ...)
}
\arguments{
\item{x}{An incidence object, generated by the function \code{\link{incidence}}.}

\item{split}{An optional time point identifying the separation between the two models. If NULL,
a single model is fitted. If provided, two models would be fitted on the time periods on either
side of the split.}

\item{level}{The confidence interval to be used for predictions; defaults to 95\%.}
}
\description{
This function fits two exponential models to incidence data, of the form:
\eqn{log(y)} = r * t + b
}
\details{
where 'y' is the incidence, 't' is time (in days), 'r' is the growth rate, and 'b' is the
origin.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the \code{\link{incidence}} function to generate the 'incidence' objects.
}

