% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_tidy.R
\name{get_info_tidy}
\alias{get_info_tidy}
\alias{tidy_incidence}
\alias{glance_incidence}
\title{tidy up incidence outputs}
\usage{
get_info_tidy(incidence_fit, what)

tidy_incidence(incidence_fit)

glance_incidence(incidence_fit)
}
\arguments{
\item{incidence_fit}{incidence_fit or incidence_fit_list object}

\item{what}{character string of parameter of interest}
}
\value{
incidence tibbles!
}
\description{
functions to get tibble outputs from incidence_fit objects
}
\section{Functions}{
\itemize{
\item \code{get_info_tidy}: generates a tibble of incidence::get_info()

\item \code{tidy_incidence}: generates a complete summary tibble from incidence fit paramteter estimates

\item \code{glance_incidence}: generates a complete summary tibble from incidence fit model performance
}}

\examples{

# example outbreak --------------------------------------------------------

library(outbreaks) #sample data
library(incidence) #core functions
# library(avallecam) #improvements
dat <- ebola_sim$linelist$date_of_onset
i.7 <- incidence(dat, interval=7)
# plot(i.7)
f1 <- fit(i.7[1:20])
# f1
f2 <- fit_optim_split(i.7)
# f2

# get_info_tidy: tibble outputs from fit object ---------------------------

# original
get_info(f1,what = "r")

# new improvement
get_info_tidy(f1,what = "r")

# for incidence::fit_optim_split outputs use .$fit

# original
get_info(f2$fit, what = "r")

#improvement
get_info_tidy(f2$fit, what = "r")

# try more examples:

# get_info(f2$fit, what = "r.conf")
# get_info_tidy(f2$fit, what = "r.conf")
# get_info_tidy(f2$fit, what = "doubling")
# get_info_tidy(f2$fit, what = "doubling.conf")
# get_info_tidy(f2$fit, what = "halving")
# get_info_tidy(f2$fit, what = "halving.conf")

# tidy_incidence: all key parameters in tibble ----------------------------

tidy_incidence(incidence_fit = f1)
tidy_incidence(incidence_fit = f2$fit)

# glance_incidence: check model performance -------------------------------

glance_incidence(incidence_fit = f1)
glance_incidence(incidence_fit = f2$fit)

}
