% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\alias{get_n.default}
\alias{get_n.incidence}
\title{Get n for an incidence object}
\usage{
get_n(x)

\method{get_n}{default}(x)

\method{get_n}{incidence}(x)
}
\arguments{
\item{x}{an \link{incidence} object}
}
\value{
an \code{integer} denoting the number of samples represented by the
incidence object.
}
\description{
Get n for an incidence object
}
\examples{
i <- incidence(sample(-3:50, 100, replace = TRUE))
get_n(i)
}
