% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timespan.R
\name{get_timespan}
\alias{get_timespan}
\alias{get_timespan.default}
\alias{get_timespan.incidence}
\title{Get timespan for an incidence object}
\usage{
get_timespan(x)

\method{get_timespan}{default}(x)

\method{get_timespan}{incidence}(x)
}
\arguments{
\item{x}{an \link{incidence} object}
}
\value{
an \code{integer} denoting the timespan represented by the incidence
object.
}
\description{
Get timespan for an incidence object
}
\examples{
i <- incidence(sample(-3:50, 100, replace = TRUE))
get_timespan(i)
}
