% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.list.R
\name{incidence.list}
\alias{incidence.list}
\title{This is the method that is used when the input into incidence() is of the class list. This method prints a warning if the list has more than one members. It then takes the first member and re-runs incidence, which will choose a method according to the class of the first member of the input.}
\usage{
\method{incidence}{list}(SymptomOnset, Interval = 1, ...)
}
\arguments{
\item{SymptomOnset}{A list with the first member of length n, where n is the number of cases}

\item{Interval}{An integer (number of days in the interval, default = 1)}
}
\value{
A named list with 2 members (Days [first day per interval] and NoOfCases [count of incidents during the respective interval])
}
\description{
This is the method that is used when the input into incidence() is of the class list. This method prints a warning if the list has more than one members. It then takes the first member and re-runs incidence, which will choose a method according to the class of the first member of the input.
}
\examples{
incidence(list(c("2016-01-01", "2016-01-05", "2016-01-08", "2016-01-03", "2016-01-07", "2016-01-02", "2016-01-04", "2016-01-06", "2016-01-09", "2016-01-02")))
incidenceComputation(list(c(1, 5, 8, 3, 7, 2, 4, 6, 9, 2)), 2)
}

