% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.numeric.R
\name{incidence.numeric}
\alias{incidence.numeric}
\title{This is the method that is used when the input into incidence() is of the
class numeric. This method shifts the values so that the first incidence is on
day 1. incidenceComputation is then used to compute the number of incidents per
interval. This result is then again shifted back to the original starting date.}
\usage{
\method{incidence}{numeric}(SymptomOnset, Interval = 1, ...)
}
\arguments{
\item{SymptomOnset}{An nx1 numerical vector, where n is the number of cases}

\item{Interval}{An integer (number of days in the interval, default = 1)}
}
\value{
Incidents A named list with 2 members (Days [first day per interval] and
   NoOfCases [count of incidents during the respective interval])
}
\description{
This is the method that is used when the input into incidence() is of the
class numeric. This method shifts the values so that the first incidence is on
day 1. incidenceComputation is then used to compute the number of incidents per
interval. This result is then again shifted back to the original starting date.
}
\examples{
incidenceComputation(c(1, 5, 8, 3, 7, 2, 4, 6, 9, 2))
incidenceComputation(c(1, 5, 8, 3, 7, 2, 4, 6, 9, 2), 2)
}

