% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{linelist2xts}
\alias{linelist2xts}
\title{Simple Date based aggregation based on the data.table function.}
\usage{
linelist2xts(formula, dateProjFun = identity, dRange = NULL, data)
}
\arguments{
\item{formula}{Formula of type LHSVar ~ RHSVar, where LHSVar is
a column of class Dates and RHSVar is a column containing a factor variable}

\item{dateProjFun}{Date projection function, i.e. how are dates projected down (i.e. monday, start of month, etc.). Result must be a Date again!}

\item{dRange}{Vector of length two giving the start and stop dates, the actual range will be range(dateProjFun(dRange)).}

\item{data}{data.frame containing the LHSVar RHSVar variables.}
}
\value{
xts time series
}
\description{
The function allows to split by a single factor variable. More factors
as well as factor interactions are currently not supported.
}
\details{
The formula has to be of the type LHSVar ~ RHSVar, where LHSVar is
a column of class Dates and RHSVar is a column containing a factor variable.
Currently, the formula is unrwapped manually. A better solution is to use
formula syntax. Note: A multivariate time series for all levels of RHSVar
(including NA if present) is generated.
}

