% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pal1}
\alias{pal1}
\alias{pal1dark}
\alias{pal1light}
\alias{palettes}
\title{Color palettes used in incidence}
\usage{
pal1(n)

pal1light(n)

pal1dark(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in incidence.
}
\examples{

plot(1:4, cex=8, pch=20, col = pal1(4), main="palette: pal1")
plot(1:100, cex=8, pch=20, col = pal1(100), main="palette: pal1")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

