% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.incidence}
\alias{plot.incidence}
\title{Plot function for incidence objects}
\usage{
\method{plot}{incidence}(x, ..., border = NA, xlab = "", ylab = NULL)
}
\arguments{
\item{x}{An incidence object, generated by the function \code{\link{incidence}}.}

\item{...}{Further arguments passed to other methods (currently not used).}

\item{border}{The color to be used for the borders of the bars; NA for invisiable borders.}

\item{xlab}{The label to be used for the x-axis; empty by default.}

\item{ylab}{The label to be used for the y-axis; by default, a label will be generated
automatically according to the time interval used in incidence computation.}
}
\description{
This function is used to visualise the output of the \code{\link{incidence}} function, using the
package \code{ggplot2}.
}
\examples{

if(require(outbreaks)) {
  onset <- ebola.sim$linelist$date.of.onset

  ## daily incidence
  inc <- incidence(onset)
  inc
  plot(inc)

  ## weekly incidence
  inc.week <- incidence(onset, interval = 7)
  inc.week
  plot(inc.week)
  plot(inc.week, border = "white") # with visible border
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the \code{\link{incidence}} function to generate the 'incidence' objects.
}

