% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.optim.split}
\alias{plot.incidence.fit}
\alias{print.incidence.fit}
\title{Fit exponential models to incidence data}
\usage{
fit(x, split = NULL, level = 0.95)

fit.optim.split(x, window = x$timespan/4, plot = TRUE)

\method{print}{incidence.fit}(x, ...)

\method{plot}{incidence.fit}(x, ..., col.pal = pal1)
}
\arguments{
\item{x}{An incidence object, generated by the function \code{\link{incidence}}. For the
plotting function, an \code{incidence.fit} object.}

\item{split}{An optional time point identifying the separation between the two models. If NULL,
a single model is fitted. If provided, two models would be fitted on the time periods on either
side of the split.}

\item{level}{The confidence interval to be used for predictions; defaults to 95\%.}

\item{window}{The size, in days, of the time window either side of the split.}

\item{plot}{A logical indicating whether a plot should be added to the output, showing the mean
R2 for various splits.}

\item{...}{further arguments passed to other methods (not used)}

\item{col.pal}{The color palette to be used for the groups; defaults to \code{pal1}. See
\code{\link{pal1}} for other palettes implemented in incidence.}
}
\description{
This function fits two exponential models to incidence data, of the form:
\eqn{log(y)} = r * t + b
}
\details{
where 'y' is the incidence, 't' is time (in days), 'r' is the growth rate, and 'b' is the
origin. The function \code{fit} will fit one model by default, but will fit two models on either side of a splitting data if the argument \code{split} is provided.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the \code{\link{incidence}} function to generate the 'incidence' objects.
}

