% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit_optim_split}
\alias{plot.incidence_fit}
\alias{print.incidence_fit}
\title{Fit exponential models to incidence data}
\usage{
fit(x, split = NULL, level = 0.95, quiet = FALSE)

fit_optim_split(x, window = x$timespan/4, plot = TRUE, quiet = TRUE)

\method{print}{incidence_fit}(x, ...)

\method{plot}{incidence_fit}(x, ..., col_pal = pal1)
}
\arguments{
\item{x}{An incidence object, generated by the function
\code{\link{incidence}}. For the plotting function, an \code{incidence_fit}
object.}

\item{split}{An optional time point identifying the separation between the
two models. If NULL, a single model is fitted. If provided, two models would
be fitted on the time periods on either side of the split.}

\item{level}{The confidence interval to be used for predictions; defaults to
95\%.}

\item{quiet}{A logical indicating if warnings from \code{fit} should be
hidden; FALSE by default. Warnings typically indicate some zero incidence,
which are removed before performing the log-linear regression.}

\item{window}{The size, in days, of the time window either side of the
split.}

\item{plot}{A logical indicating whether a plot should be added to the
output, showing the mean R2 for various splits.}

\item{...}{further arguments passed to other methods (not used)}

\item{col_pal}{The color palette to be used for the groups; defaults to \code{pal1}. See
\code{\link{pal1}} for other palettes implemented in incidence.}
}
\value{
For \code{fit}, a list with the class \code{incidence_fit} (for a
single model), or a list containing two \code{incidence_fit} objects (when
fitting two models). \code{incidence_fit} objects contain:

\itemize{
 \item lm: the fitted linear model

 \item info: a list containing various information extracted from the model
(detailed further)

 \item origin: the date corresponding to day '0'
}

The \code{$info} item is a list containing:

\itemize{
 \item r: the growth rate

 \item r.conf: the confidence interval of 'r'

 \item pred: a \code{data.frame} containing predictions of the model,
including the true dates (\code{dates}), their numeric version used in the
model (\code{dates.x}), the predicted value (\code{fit}), and the lower
(\code{lwr}) and upper (\code{upr}) bounds of the associated confidence
interval.

 \item doubling: the predicted doubling time in days; exists only if 'r' is
positive

 \item doubling.conf: the confidence interval of the doubling time

 \item halving: the predicted halving time in days; exists only if 'r' is
negative

 \item halving.conf: the confidence interval of the halving time
}

For \code{fit_optim_split}, a list containing:
\itemize{

 \item df: a \code{data.frame} of dates that were used in the optimization
procedure, and the corresponding average R2 of the resulting models.

 \item split: the optimal splitting date

 \item fit: the resulting \code{incidence_fit} objects

 \item plot: a plot showing the content of \code{df} (ggplot2 object)
}
}
\description{
The function \code{fit} fits two exponential models to incidence data, of
the form: \eqn{log(y) = r * t + b} \cr
where 'y' is the incidence, 't' is time (in days), 'r' is the growth rate,
and 'b' is the origin. The function \code{fit} will fit one model by
default, but will fit two models on either side of a splitting date
(typically the peak of the epidemic) if the argument \code{split} is
provided. The function \code{fit_optim_split} can be used to find the
optimal 'splitting' date, defined as the one for which the best average R2
of the two models is obtained.
}
\examples{

if (require(outbreaks)) {
  dat <- ebola.sim$linelist$date.of.onset

 ## EXAMPLE WITH A SINGLE MODEL

  ## compute weekly incidence
  i.7 <- incidence(dat, interval=7)
  plot(i.7)
  plot(i.7[1:20])

  ## fit a model on the first 20 weeks
  f <- fit(i.7[1:20])
  f
  names(f)
  head(f$pred)

  ## plot model alone (not recommended)
  plot(f)

 ## plot data and model (recommended)
 plot(i.7, fit=f)
 plot(i.7[1:25], fit=f)


## EXAMPLE WITH 2 PHASES
 ## specifying the peak manually
 f2 <- fit(i.7, split=as.Date("2014-10-15"))
 f2
 plot(i.7, fit=f2)

## finding the best 'peak' date
f3 <- fit_optim_split(i.7)
f3
plot(i.7, fit=f3$fit)
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the \code{\link{incidence}} function to generate the 'incidence'
objects.
}

