% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{get_fit}
\alias{get_fit}
\alias{get_fit.incidence_fit}
\alias{get_fit.incidence_fit_list}
\title{Accessors for \code{incidence_fit} objects}
\usage{
get_fit(x)

\method{get_fit}{incidence_fit}(x)

\method{get_fit}{incidence_fit_list}(x)
}
\arguments{
\item{x}{an \code{incidence_fit} or \code{incidence_fit_list}
object.}
}
\value{
a list of \code{incidence_fit} objects.
}
\description{
Accessors for \code{incidence_fit} objects
}
\examples{

if (require(outbreaks)) {
 dat <- ebola_sim$linelist$date_of_onset

 ## EXAMPLE WITH A SINGLE MODEL

 ## compute weekly incidence
 sex <- ebola_sim$linelist$gender
 i.sex <- incidence(dat, interval = 7, group = sex)
 
 ## Compute the optimal split for each group separately
 fits  <- fit_optim_split(i.sex, separate_split = TRUE)

 ## `fits` contains an `incidence_fit_list` object
 fits$fit
 
 ## Grab the list of `incidence_fit` objects
 get_fit(fits$fit)
 
}
}
