% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{incidence_pal1}
\alias{incidence_pal1}
\alias{palettes}
\alias{incidence_pal1_light}
\alias{incidence_pal1_dark}
\title{Color palettes used in incidence}
\usage{
incidence_pal1(n)

incidence_pal1_light(n)

incidence_pal1_dark(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in incidence.
}
\examples{

plot(1:4, cex=8, pch=20, col = incidence_pal1(4),
     main = "palette: incidence_pal1")
plot(1:100, cex=8, pch=20, col = incidence_pal1(100),
     main ="palette: incidence_pal1")
plot(1:100, cex=8, pch=20, col = incidence_pal1_light(100),
     main="palette: incidence_pal1_light")
plot(1:100, cex=8, pch=20, col = incidence_pal1_dark(100),
     main="palette: incidence_pal1_dark")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
