% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.incidence}
\alias{plot.incidence}
\title{Plot function for incidence objects}
\usage{
\method{plot}{incidence}(x, ..., fit = NULL, stack = is.null(fit),
  color = "black", border = NA, col_pal = pal1, alpha = 0.7,
  xlab = "", ylab = NULL)
}
\arguments{
\item{x}{An incidence object, generated by the function \code{\link{incidence}}.}

\item{...}{Further arguments passed to other methods (currently not used).}

\item{fit}{An 'incidence_fit' objet as returned by \code{\link{fit}}.}

\item{stack}{A logical indicating if bars of multiple groups should be stacked, or displayed
side-by-side.}

\item{color}{The color to be used for the filling of the bars; NA for invisiable bars; defaults to "black".}

\item{border}{The color to be used for the borders of the bars; NA for invisiable borders; defaults to NA.}

\item{col_pal}{The color palette to be used for the groups; defaults to \code{pal1}. See
\code{\link{pal1}} for other palettes implemented in incidence.}

\item{alpha}{The alpha level for color transparency, with 1 being fully opaque and 0 fully
transparent; defaults to 0.7.}

\item{xlab}{The label to be used for the x-axis; empty by default.}

\item{ylab}{The label to be used for the y-axis; by default, a label will be generated
automatically according to the time interval used in incidence computation.}
}
\description{
This function is used to visualise the output of the \code{\link{incidence}} function, using the
package \code{ggplot2}.
}
\examples{

if(require(outbreaks)) {
  onset <- ebola.sim$linelist$date.of.onset

  ## daily incidence
  inc <- incidence(onset)
  inc
  plot(inc)

  ## weekly incidence
  inc.week <- incidence(onset, interval = 7)
  inc.week
  plot(inc.week)
  plot(inc.week, border = "white") # with visible border

  ## use group information
  sex <- ebola.sim$linelist$gender
  inc.week.gender <- incidence(onset, interval = 7, groups = sex)
  plot(inc.week.gender)

  ## adding fit
  fit <- fit_optim_split(inc.week.gender)$fit
  plot(inc.week.gender, fit = fit)
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
the \code{\link{incidence}} function to generate the 'incidence' objects.
}

